abstract away compiler function
