+++ playwright/utils/provider_setup.ts
@@ -1,50 +1,13 @@
-import webpack from 'webpack'
 import path from 'path'
-import fs from 'fs'
 import Web3EthAccounts, { Accounts as AccountsType } from 'web3-eth-accounts'
 
 import { context } from './test_setup'
+import { compileInjects, fileExists } from './build_injects'
 
 // get connection data from CONFIG
 const { config } = CONFIG.defaultProviderConfig
 const providerURL = 'ethNodeUrl' in config ? config.ethNodeUrl : 'https://rinkeby.infura.io/v3/' + config.infuraId
 console.log('url', providerURL)
-
-const fileExists = (path: string): Promise<boolean> => {
-  return fs.promises.access(path).then(
-    () => true,
-    () => false,
-  )
-}
-
-const compileInjects = async (): Promise<void> => {
-  // only rebuild when no file found
-  // or when passed PWREBUILD=1 env var
-  if (!process.env.PWREBUILD && (await fileExists(path.resolve(__dirname, '../build/inject_provider.js')))) {
-    return
-  }
-
-  console.log('path.resolve', path.resolve(__dirname, '../build/'))
-  const compiler = webpack({
-    entry: path.resolve(__dirname, '../inject_provider.js'),
-    output: {
-      path: path.resolve(__dirname, '../build/'),
-      filename: 'inject_provider.js',
-    },
-    cache: true,
-  })
-
-  await new Promise((resolve, reject) => {
-    compiler.run((error, stats) => {
-      console.log('error', error)
-      if (error) {
-        return reject(error)
-      }
-
-      resolve(stats)
-    })
-  })
-}
 
 declare global {
   interface Window {
@@ -60,7 +23,11 @@ export const account = accountCreator.create()
 
 beforeAll(async () => {
   console.log('PROVIDER SETUP')
-  await compileInjects()
+  // only rebuild when no file found
+  // or when passed PWREBUILD=1 env var
+  if (process.env.PWREBUILD || !(await fileExists(path.resolve(__dirname, '../build/inject_provider.js')))) {
+    await compileInjects()
+  }
 
   await context.addInitScript({
     path: path.resolve(__dirname, '../build/inject_provider.js'),
